+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/app.ts
@@ -64,4 +64,8 @@ export class App {
         }
     }
 
+    get panelHeight():number {
+        return window.innerHeight - 20;
+    }
+
 }