+++ fiddles/angular2/fiddle-0030-D3VerticalBarChart/src/bar-chart/bar-chart.ts
@@ -141,7 +141,4 @@ export class BarChart {
         }
     }
 
-
-
-
 }