+++ fiddles/angular2/fiddle-0033-Tooltip/config.js
@@ -1,0 +1,15 @@
+System.config({
+    transpiler: 'typescript',
+    typescriptOptions: {
+        emitDecoratorMetadata: true
+    },
+    map: {
+        app: "./src"
+    },
+    packages: {
+        app: {
+            main: './boot.ts',
+            defaultExtension: 'ts'
+        }
+    }
+});