+++ fiddles/angular2/fiddle-0033-Tooltip/src/tooltip/tooltip.interface.ts
@@ -1,0 +1,11 @@
+import {Injectable} from 'angular2/core';
+import Injectable = ng.Injectable;
+
+
+@Injectable()
+export interface TooltipInterface {
+    top?:number;                                    // absolute top position of the tooltip
+    left?:number;                                   // absolute left position of the tooltip
+    lines?:Array<{ field:string, value:string}>;    // field/value collection
+    visible?:boolean;                               // visibility (hide/show) flag
+}